+++ ecms_base/modules/custom/ecms_workflow/tests/src/Unit/EcmsWorkflowBundleCreateTest.php
@@ -208,39 +208,10 @@ class EcmsWorkflowBundleCreateTest extends UnitTestCase {
       ->with(["id" => self::WORKFLOW_ID])
       ->willReturn([$workflow]);
 
-    $nodeTypeEntity = $this->createMock(ConfigEntityInterface::class);
-    $nodeTypeEntity->expects($this->exactly(12))
-      ->method('setThirdPartySetting')
-      ->withConsecutive(
-        ['scheduler', 'expand_fieldset', 'when_required'],
-        ['scheduler', 'fields_display_mode', 'fieldset'],
-        ['scheduler', 'publish_enable', 1],
-        ['scheduler', 'publish_past_date', 0],
-        ['scheduler', 'publish_past_date_created', 'error'],
-        ['scheduler', 'publish_required', 0],
-        ['scheduler', 'publish_revision', 0],
-        ['scheduler', 'publish_touch', 0],
-        ['scheduler', 'show_message_after_update', 1],
-        ['scheduler', 'unpublish_enable', 1],
-        ['scheduler', 'unpublish_required', 0],
-        ['scheduler', 'unpublish_revision', 0]
-      )
-      ->willReturnSelf();
-
-    $nodeTypeEntity->expects($this->once())
-      ->method('save')
-      ->willReturnSelf();
-
-    $nodeTypeStorage = $this->createMock(EntityStorageInterface::class);
-    $nodeTypeStorage->expects($this->once())
-      ->method('load')
-      ->with($contentType)
-      ->willReturn($nodeTypeEntity);
-
-    $this->entityTypeManager->expects($this->exactly(3))
+    $this->entityTypeManager->expects($this->exactly(2))
       ->method('getStorage')
-      ->withConsecutive(['user_role'], ['node_type'], ['workflow'])
-      ->willReturnOnConsecutiveCalls($this->entityStorage, $nodeTypeStorage, $workflowStorage);
+      ->withConsecutive(['user_role'], ['workflow'])
+      ->willReturnOnConsecutiveCalls($this->entityStorage, $workflowStorage);
 
     $testClass = new EcmsWorkflowBundleCreate($this->entityTypeManager);